Separated out data to Service and Controller
Ensured model data and controller logic is separated into Controllers
and Services.

A dude list of cities with some weather data is looped through and
displayed and cities can be added to the list.

Still need to ensure the list refreshes when a city is added and neaten
up UI.